refactor(workers): reorganize functions
